added functionality for a push notification to occur, but the part where the notification page is created is still broken(so it's difficult to test right now)
