+++ RemindIQ/RemindIQ/Views/MainPage.xaml
@@ -44,14 +44,7 @@
             </ListView.ItemTemplate>
         </ListView>
         <!-- Place new controls here -->
-        <Label Text="Welcome to Xamarin.Forms!" 
-           HorizontalOptions="Center"
-           VerticalOptions="CenterAndExpand" />
-        <Button
-            Clicked="Button_Clicked"
-            HorizontalOptions="Center"
-            Text="Send Local Notification"
-            VerticalOptions="CenterAndExpand" />
+        
     </StackLayout>
 
 </ContentPage>
