+++ RemindIQ/RemindIQ/Views/MainPage.xaml.cs
@@ -102,61 +102,5 @@ namespace RemindIQ.Views
             Reminder reminder = (Reminder)e.SelectedItem;
             await Navigation.PushModalAsync(new NavigationPage(new ReminderPage(reminder)));
         }
-
-        //the sample used a single button to generate the notification, so obviously this needs to go somewhere else
-        private void Button_Clicked(object sender, EventArgs e)
-        {
-            
-            _count++;//the sample kept track of how many times the button was pressed, so this could probably be removed no problem
-
-            var list = new List<string>
-            {
-                typeof(NotificationPage).FullName,
-                _count.ToString()
-            };
-
-            var serializer = new ObjectSerializer<List<string>>();
-            var serializeReturningData = serializer.SerializeObject(list);
-
-            var request = new NotificationRequest//the notification object that will be shown to the user in the end
-            {
-                NotificationId = 100,//think this has to match the channel id
-                Title = "Test",//reminder name should go here
-                Description = $"Tap Count: {_count}",//reminder descriptions should go here
-                BadgeNumber = _count, //not sure about this one
-                ReturningData = serializeReturningData,
-                Android =
-                {
-                    //IconName = "my_icon",
-                    //AutoCancel = false,
-                    //Ongoing = true//not really sure what this boolean does
-                },
-            };
-
-            // if not specified, default sound will be played.
-            /*
-            if (CustomSoundSwitch.IsToggled)
-            {
-                request.Sound = Device.RuntimePlatform == Device.Android
-                    ? "good_things_happen"
-                    : "good_things_happen.aiff";
-            }
-            */
-
-            // if not specified, notification will show immediately.
-            /*
-            if (UseNotifyTimeSwitch.IsToggled)
-            {
-                var notifyDateTime = NotifyDatePicker.Date.Add(NotifyTimePicker.Time);
-                if (notifyDateTime <= DateTime.Now)
-                {
-                    notifyDateTime = DateTime.Now.AddSeconds(10);
-                }
-                request.NotifyTime = notifyDateTime;
-            }
-            */
-
-            NotificationCenter.Current.Show(request);//sends object to the notification center, which is in the Android/iOS code
-        }
     }
 }
