+++ RemindIQ/RemindIQ/Views/NotificationPage.xaml.cs
@@ -6,17 +6,18 @@ using System.Threading.Tasks;
 
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
+using RemindIQ.Models;
 
 namespace RemindIQ.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class NotificationPage : ContentView
     {
-        public NotificationPage(int count)
+        public NotificationPage(string rem)
         {
             InitializeComponent();
 
-            TapCountLabel.Text = $"Tap count {count}";
+            TapCountLabel.Text = $"test of location '{rem}' being within range";
         }
     }
 }
